Shuffle the code to avoid ImportError, as per suggestion from @MarekBleschke
